syscall: Rename numbers.rs to def.rs
Rename numbers.rs to def.rs so that this file can be extended in the
future to contain more definitions, like MMFlag, syscall error code
which are common for both kernel and user.
